% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishblicc_Create.R
\name{BH_recruitment}
\alias{BH_recruitment}
\title{Creates a Beverton & Holt recruitment function}
\usage{
BH_recruitment(lR0, Rb, lRs, Rac = 0)
}
\arguments{
\item{lR0}{Log maximum recruitment}

\item{Rb}{Second BH parameter defining an inflection point}

\item{lRs}{Log-recruitment normal standard deviation}

\item{Rac}{Log-recruitment autocorrelation coefficient}
}
\value{
A function taking 2 parameters, SSB and lRd - a vector of 2 log
recruitment deviates
}
\description{
A function is produced that given a SSB and recruitment process error
calculates the corresponding recruitment. The model allows for
autocorrelated errors.
}
\details{
Expected recruitment R = R_0 * S / (R_b + S)

The function produced takes a SSB vector of length equal to the number of
simulations and a lRd vector of 2 log recruitment deviates, the second being
the current log-deviate. This allows autocorrelation to be included.
}
