% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCRPerform.R
\name{combine_HCR_results}
\alias{combine_HCR_results}
\title{JABBAstan Combines multiple HCR simulation results into a single tibble}
\usage{
combine_HCR_results(HCR_res_list, HCR_MSE)
}
\arguments{
\item{HCR_res_list}{A list of results from \link{HCR_performance}}

\item{HCR_MSE}{HCR function produced by \link{create_fishblicc_MSE},
\link{create_JABBA_MSE} or \link{create_ptStan_MSE}}
}
\value{
A tibble containing combined results from the simulations
}
\description{
JABBAstan Combines multiple HCR simulation results into a single tibble
}
