% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JBH_Create.R
\name{CPUE_MSY}
\alias{CPUE_MSY}
\title{Gets the median CPUE expected at MSY from a JABBA fit}
\usage{
CPUE_MSY(jabba_fit)
}
\arguments{
\item{jabba_fit}{A list returned from package JABBA \code{fit_jabba} function}
}
\value{
A 1D array of each CPUE at MSY
}
\description{
Gets the median CPUE expected at MSY from a JABBA fit
}
