% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JBH_Create.R
\name{create_empirical_index}
\alias{create_empirical_index}
\title{Create the empirical abundance index update function.}
\usage{
create_empirical_index(ma, func = NULL, se)
}
\arguments{
\item{ma}{Moving average parameter for index time series}

\item{func}{Function used to combine multiple CPUE series into a single index}

\item{se}{Observation standard error to be added in to the index}
}
\value{
A function taking a single CPUE list as a parameter that can be used
to calculate the HCR index
}
\description{
The index function created calculates a moving average from one or more
abundance indices, usually CPUE.
}
\details{
The same lognormal error is applied across all simulations to make HCR more
comparable.
}
