% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishblicc_Create.R
\name{create_fishblicc_MSE}
\alias{create_fishblicc_MSE}
\title{Create a HCR MSE function based on a fishblicc fit.}
\usage{
create_fishblicc_MSE(
  fishblicc_fit,
  vbK,
  catch_weight = 1000,
  control_type = "Effort",
  index_type = "MeanLength",
  nsim = 1000,
  proj_length = 50,
  StartYear = 1,
  AnnualRecruitment = TRUE,
  yr_steps = 0,
  AssessmentPeriod = yr_steps,
  SampleSize = 0,
  SRDelay = 1,
  h_steepness = 0.75,
  autocorr = 0,
  recruit_cv = 0.3,
  sim_recruit = "independent",
  ref_control = NULL,
  rp_type = "SSB0",
  rseed = Sys.time()
)
}
\arguments{
\item{fishblicc_fit}{A list returned from package fishblicc \code{blicc_ref_pts}
function (required)}

\item{vbK}{The von Bertalanfy growth rate parameter K (/year). Can be a
single value, or a vector the same length as nsim to allow for random
variation. (required)}

\item{catch_weight}{The total catch weight for the year fishblicc was fit
(last data year). Used for scaling values so catch values align in
magnitude. (default=1000)}

\item{control_type}{Type of control applied by HCR. Only "Effort" is
currently supported for this model.}

\item{index_type}{HCR index used (default="MeanLength")}

\item{nsim}{Number of simulations to run (default=1000)}

\item{proj_length}{Projection length in years (default=50)}

\item{StartYear}{Year projections start, used for plot labels only (default=1).}

\item{AnnualRecruitment}{= TRUE,}

\item{yr_steps}{Time steps within each year. The higher number is more
accurate but slower simulations. Can be left to default so each step is
equivalent to max(2, vbK=0.2) (default=0, i.e. estimate).}

\item{AssessmentPeriod}{Period within the year when the HCR is calculated for
the following year's control (default=yr_steps).}

\item{SampleSize}{The length frequency sample size. If not provided, the
fishblicc sample size is used  (default=0, i.e. estimate).}

\item{SRDelay}{annual delay between SSB calculated subsequent recruitment
(default=1).}

\item{h_steepness}{Beverton and Holt stock recruitment relationship steepness
parameter (default=0.75).}

\item{autocorr}{Autocorrelation (0-1.0) for recruitment deviates (default=0).}

\item{recruit_cv}{Lognormal sigma (coefficient of variation) for recruitment
deviates (default=0.3)}

\item{sim_recruit}{recruitment deviates can be 'shared' across the simulated
parameters set or generated 'independent'-ly for each parameter set.}

\item{ref_control}{A reference control (fishing mortality) used to multiply
the HCR control. This is usually the current fishing mortality, so HCR
changes are relative to the current fishery (default = NULL i.e. current
fishing mortality estimated in fishblicc).}

\item{rp_type}{= Either "SSB0", "MSY" or current (default = "SSB0").
See details.}

\item{rseed}{Random seed if needed for repeated simulations (default = Sys.time()).
This random seed is used for each HCR tested using the returned
function.}
}
\value{
A function that will apply the HCR using the fishblicc model fit and
specified HCR parameters: index-control inflection point vectors, the
control change limit and moving average parameter. See details.
}
\description{
The HCR function returns simulation results from HCR parameters using a
fishblicc model fit results.
}
\details{
The the same state space lognormal random number can be applied across
simulations, so HCR performance is more comparible.

Uses fishblicc results from the \code{blicc_ref_pts} function for growth and
mortality parameters, but rate parameter K, and recruitment parameters must
be provided to convert fishblicc from a static model to a time series
model. The MCMC draws are used to provide a measure of uncertainty from
the posterior fit.

The model is a length and age structured simulation of a single species.
Growth variability is simulated using 50 growth-type-groups over a
Gamma-distributed Linf parameter based on the fishblicc fit. The supplied
von Bertalanffy growth rate (K) can also be randomly drawn. Most parameters
apart from K are taken from the fishblicc fit.

The stock recruitment relationship is the Beverton and Holt function, with
steepness set by the user. A default precautionary steepness = 0.75 is
applied This is raised if the population dynamics do not support this
steepness level and a warning is given. This can occur when the estimated
SPR is very low. The unexploited recruitment is estimated as scaling factor
consistent with the catch (catch_weight).

Reference points are estimated. The default reference point is 40\%
unexploited SSB, which is precautionary and consistent with 40\% SPR. However,
SPR does not take account of a stock recruitment relationship which is
included in this model. MSY reference points are also provided for reference.
These can be used for the target, but in general this is not recommended in
data-poor situations as estimated values depend entirely on the assumed
stock-recruitment and are generally not precautionary. You could treat MSY as
a limit reference point and set the target as two times MSY, which could work
well in some circumstances (see \link{fb_set_ref_pt_type}).

The HCR function that is produced takes six parameters.
\itemize{
\item trIndex: a vector of HCR index inflection points
\item trControl: a single vector or list of vectors, one for each gear
\item change_limit: an annual proportional change limit on the control, if any
\item ma: a moving average parameter for the index (default=0.5)
\item control_type: ignored (default="Effort")
\item ctrl_pF: Proportion of the fishing mortality affected by the control. So
(1-ctrl_pF)*F is fixed and unaffected. This could represent uncontrolled
subsistence fishing for example (default=1).
}
}
