% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JABBA_Create.R
\name{create_JABBA_MSE}
\alias{create_JABBA_MSE}
\title{Create a HCR MSE function}
\usage{
create_JABBA_MSE(
  jabba_fit,
  proj_length = 50,
  nsim = 1000,
  ref_pt = standard_risk_ref_pt()
)
}
\arguments{
\item{jabba_fit}{A list returned from package JABBA \code{fit_jabba} function}

\item{proj_length}{Projection length in years}

\item{nsim}{Number of simulations to run}

\item{ref_pt}{Specifies risk based reference points in a list. Otherwise
defaults apply.}

\item{control_type}{Type of control applied by HCR either "Effort" or "Catch"}
}
\value{
A function that will apply the HCR using the JABBA model fit and
specified HCR parameters: index-control inflection point vectors, the
control change limit and moving average parameter.
}
\description{
The HCR function returns simulation results from HCR parameters using a
JABBA model fit results.
}
\details{
The the same state space lognormal random number is applied across
simulations, so HCR performance is more comparible.
}
