% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JBH_Create.R
\name{create_linear_control}
\alias{create_linear_control}
\title{Calculate a linear sliding control to be applied based on the index value}
\usage{
create_linear_control(trIndex, trControl, change_limit = NA, control_type)
}
\arguments{
\item{trIndex}{A sorted list of index inflexion points from low to high.
Same length as \code{trControl}}

\item{trControl}{A sorted list of control inflection points defining linear
HCR.}

\item{change_limit}{An annual limit on the control change as a proportion}

\item{control_type}{Either "Effort" or "Catch"}
}
\value{
A function to calculate the new control from the index and previous
year's control
}
\description{
The function creates a function to calculate a control based on the index,
previous control and HCR parameters held in the HCR list.
}
