% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JABBAstan_Create.R
\name{create_ptStan_MSE}
\alias{create_ptStan_MSE}
\title{Create a HCR MSE function using a Stan fit of the JABBA}
\usage{
create_ptStan_MSE(
  stan_fit,
  stan_dat,
  proj_length = 50,
  nsim = 1000,
  implement_cv = "estimate",
  ref_pt = standard_risk_ref_pt()
)
}
\arguments{
\item{stan_fit}{A stanfit object}

\item{stan_dat}{Data list used to fit the Stan model}

\item{proj_length}{Projection length in years. If set to zero this will apply
the HCR to the historical fit using estimated errors.}

\item{nsim}{Number of simulations to run}

\item{implement_cv}{Coefficient of variation for lognormal implementation
error applied to fishing mortality. Not used if proj_length==0.}

\item{ref_pt}{Specifies risk based reference points in a list. Otherwise
defaults apply.}

\item{control_type}{Type of control applied by HCR either "Effort" or "Catch"}
}
\value{
A function that will apply the HCR using the stan model fit and
specified HCR parameters: index-control inflection point vectors, the
control change limit and moving average parameter.
}
\description{
The HCR function returns simulation results from HCR parameters using a
JABBA model fit results from JABBAstan.
}
