% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineHCR.R
\name{define_HCR_test_range}
\alias{define_HCR_test_range}
\title{A tibble is created defining full combinations of the defined parameter
ranges.}
\usage{
define_HCR_test_range(
  IMSY,
  fMSY,
  control_type,
  rel_index_range = c(0.5, 1),
  rel_control_range = c(0.1, 1),
  NInflex = 2L,
  NBreaks = 5L,
  change_limit = NA,
  ma = 0.5,
  ctrl_pF = 1
)
}
\arguments{
\item{IMSY}{Index at MSY estimated from the JABBA model}

\item{fMSY}{Control (TAC or effort) at MSY estimated from the JABBA model}

\item{rel_index_range}{(lower, upper) range for indices as proportions of
the MSY values}

\item{rel_control_range}{(lower, upper) range for controls as proportions
of the MSY values}

\item{NInflex}{The number of inflexion (trigger) points for the linear HCR}

\item{NBreaks}{The number of breaks including the ranges to generate the
test HCR}

\item{change_limit}{Limit on the annual change of the control. NA implies
no limit}

\item{ma}{Moving average parameter for index time series}

\item{ctrl_pF}{Effectiveness of the control (what proportion the harvest
activity is adjusted)}
}
\value{
A tibble of all combinations of index, control, control change limit,
and moving average parameters for the HCR.
}
\description{
Generates combinations of index, controls, change limits and the index moving
average parameter as list columns in a tibble, each row being a unique HCR.
Only works with single controls.
}
