% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineHCR.R
\name{define_inflection_points}
\alias{define_inflection_points}
\title{Generates infection points for index/control pairs of values for a list of
controls being applied that make up HCR}
\usage{
define_inflection_points(
  vInd,
  vCon,
  rel_index_range = list(c(0.5, 1.5)),
  rel_control_range = list(c(0.5, 1.5)),
  NBreaks = 5
)
}
\arguments{
\item{vInd}{A vector of index values to apply}

\item{vCon}{A vector of control values to apply}

\item{rel_index_range}{A list of index ranges relative to vInd}

\item{rel_control_range}{A list of control ranges relative to vCon}

\item{NBreaks}{The number of inflection points chosen for each line}
}
\description{
Generates infection points for index/control pairs of values for a list of
controls being applied that make up HCR
}
