% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishblicc_Create.R
\name{fb_create_control_func}
\alias{fb_create_control_func}
\title{Create a control function to be applied based on the index value above}
\usage{
fb_create_control_func(refControl, trIndex, trControl, change_limit = NA)
}
\arguments{
\item{refControl}{The F matrix with rows the number of simulations and
columns the number of gears.}

\item{trIndex}{A single sorted vector of index inflection points from low to
high. Same rows as \code{trControl}}

\item{trControl}{A matrix (or vector) of sorted vectors of control
inflection points defining linear HCR. If there is no separate column for
each gear, the vector is repeated.}

\item{change_limit}{An annual limit on the control changes as a proportion}

\item{control_type}{Must be "Effort". For future use.}
}
\value{
A function to calculate the new controls as a matrix from the index
and previous year's control. The matrix is number of simulations * the
number of gears.
}
\description{
This function returns a function to calculate a control based on the index,
The only control currently supported
is an effort control. However, the function supports controls on multiple
gears.
}
\details{
The trControl must be a matrix with a column for each gear
}
