% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishblicc_Create.R
\name{fb_create_HCR_index_func}
\alias{fb_create_HCR_index_func}
\title{Create the function to update the HCR index}
\usage{
fb_create_HCR_index_func(ma, index_func, default_factor = 1.05)
}
\arguments{
\item{ma}{Moving average parameter for index time series}

\item{index_func}{An index function that takes a single 3d length
data matrices \code{lenfq} with dimensions length, sims, gears}

\item{default_factor}{A default factor to allow recovery with no valid index
(defaults to 1.05 = 5\% increase)}
}
\value{
A function taking the previous index and a 3d length frequency matrix that can be
used to calculate the HCR index
}
\description{
The HCR indicator function that is created calculates an indicator on
stock status that can be compared to reference points. The indicator
is either a function using length frequency. An optional moving average is
applied to the index.
}
\details{
Either a mean length or length-based catch curve indicator is used.
}
