% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishblicc_Create.R
\name{fb_set_ref_pt_type}
\alias{fb_set_ref_pt_type}
\title{Set the reference point type in the HCR function}
\usage{
fb_set_ref_pt_type(HCR_MSE, rp_type, TRP, LRP = NULL)
}
\arguments{
\item{HCR_MSE}{HCR_MSE function created by \link{create_fishblicc_MSE}}

\item{rp_type}{Reference point type ("SSB0", "MSY", "Current")}

\item{TRP}{Target reference point relative to rp_type}

\item{LRP}{Limit reference point relative to rp_type}
}
\description{
Reference point types can be "SSB0" (relative to the unexploited state),
"MSY" (relative to SSB at MSY) or "Current" (relative to the values estimated
in the stock assessment). MSY is generally not recommended in data poor
situations as it will not be well estimated and can lead to risky advice.
SSB0 with the target set at precautionary values is probably a better proxy
for MSY. The "Current" setting may be useful to guide HCR that improve on the
current status where there is little confidence in what that status is.
}
