% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishblicc_Create.R
\name{fixed_recruitment}
\alias{fixed_recruitment}
\title{Creates a fixed recruitment function}
\usage{
fixed_recruitment(lR0, lRs, Rac = 0)
}
\arguments{
\item{lR0}{Log maximum recruitment}

\item{lRs}{Log-recruitment normal standard deviation}

\item{Rac}{Log-recruitment autocorrelation coefficient}
}
\description{
A function is produced that given recruitment process error
calculates the corresponding recruitment. The model allows for
autocorrelated errors. The SSB is ignored.
}
