% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineHCR.R
\name{generate_HCR_range}
\alias{generate_HCR_range}
\title{Generates valid combinations of values corresponding to index or controls}
\usage{
generate_HCR_range(values = c(0.5, 1, 1.5), NInflex = 2)
}
\arguments{
\item{values}{A vector of alternative controls or indices relative to some
value, so 1.0 would be that value}

\item{NInfex}{The number of inflection points chosen for each line}
}
\description{
The function returns a list of values with inflection points forming
straight lines, with all valid different combinations of the values vector.
}
