% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCRPerform.R
\name{graph_HCR_performance}
\alias{graph_HCR_performance}
\title{Plot the harvest control rule performance indicators}
\usage{
graph_HCR_performance(
  HCR_df,
  PIX,
  PIY,
  ref_pt = TRUE,
  eval = FALSE,
  avg_line = FALSE
)
}
\arguments{
\item{HCR_df}{A data frame produced by eg \link{define_HCR_test_range} with
HCR parameters}

\item{PIX}{Performance indicator to go on the X axis}

\item{PIY}{Performance indicator to go on the Y axis}

\item{ref_pt}{Whether to plot the performance reference points as
vertical/horizontal dotted lines}

\item{eval}{Whether to plot the evaluation for candidate vs rejected HCR}

\item{avg_line}{whether to include a line regression through the performance
indicators}
}
\value{
An XY point plot of the HCR performance indicators
}
\description{
An XY plot is returned with the requested performance indicators.
}
\details{
If requested (eval==TRUE) candidate HCR are plotted as black points and
rejected HCR as red points.

A linear fit is shown for plots covering catches and CPUE. The line goes
through the origin indicating the exchange between increasing
catch with increasing catch range. HCR above or below this line may indicate
a better "exchange rate" between measures where they conflict.

If requested (ref_pt==TRUE) and available, vertical
and horizontal dotted lines represent the reference points for each
indicator. Reference points are either the performance actually achieved
historically (proj_length=0 for the production models only) or based on the
average performance across HCR being examined (all catch based performance
indicators), the CPUE at MSY for the CPUE_Avg indicator, or the risk-based
reference points set in the projection function for the stock state
indicators and error type indicators.

The variables that can be plotted are:
Catch_Avg : Average catch
Catch_Rng : Catch range
Catch_pcile : Lower catch 10 percentile
CPUE_Avg : Average CPUE
lt_Blim : Proportion of the simulations below the limit reference point
at_Btar : Proportion of the simulations in the target region
State : at_Btar - lt_Blim
gt_Btar : Greater than the target region
Err_Type1 : Proportion of the simulations with a false positive
Err_Type2 : Proportion of the simulations with a false negative

The reference points and lower catch 10 percentile can be adjusted in the
model's reference point list.

For the proportion of Type 1 (false positive), the decision rule results in
management intervention reducing harvest when the stock in reality is at or
above its target level. For Type 2 (false negative), the decision rule does
not harvest when the stock is in reality below the limit reference point,
which in this case is considered worse as it is less precautionary. Higher
error rates indicate poorer HCR performance and might be improved by
adjusting the index trigger points.
}
