% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JBH_HCRTest.R
\name{graph_HCR_status_catches}
\alias{graph_HCR_status_catches}
\title{Plot the HCR performance with respect to stock status and catch

A ggplot is returned with the the proportion of the simulation time across
all runs for an HCR stock status (x-axis) and lower 10\%_ile catch (y-axis).
HCR change limits indicated by colour.}
\usage{
graph_HCR_status_catches(HCR_df)
}
\arguments{
\item{HCR_df}{A data frame produced by eg \link{define_HCR_test_range} with
HCR parameters}
}
\value{
A ggplot object plotting the HCR performance catches
}
\description{
Plot the HCR performance with respect to stock status and catch

A ggplot is returned with the the proportion of the simulation time across
all runs for an HCR stock status (x-axis) and lower 10\%_ile catch (y-axis).
HCR change limits indicated by colour.
}
