% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineHCR.R
\name{graph_linear_HCR}
\alias{graph_linear_HCR}
\title{Plot harvest control rules with the stock status index on the x-axis and
control on the y-axis.}
\usage{
graph_linear_HCR(
  HCR_df,
  ctrl_index = 1L,
  HCR_sim = NULL,
  highlight = 0,
  HCR_ID = TRUE
)
}
\arguments{
\item{HCR_df}{A data frame produced by eg \link{define_HCR_test_range} with
HCR parameters}

\item{ctrl_index}{Set the control to plot. Only relevant if more than one
control is applied}

\item{HCR_sim}{Fishery HCR simulation result list from the MSE
(see \link{create_fishblicc_MSE}, \link{create_JABBA_MSE}, \link{create_ptStan_MSE}).}

\item{highlight}{A vector of HCR ID to highlight in the plot in red. The
default 0 means it is not used.}

\item{HCR_ID}{Logical - whether to include the ID as a factor in the plot.
Will only work for 12 or fewer HCR.}
}
\value{
A ggplot object plotting the HCR's
}
\description{
A ggplot is returned with the HCR plotted as a line between
the index of stock status and control. Copes with multiple HCR and multiple
gears, but plots can get overwhelming so use with care.
}
\details{
The number of gears is indicated by the number of columns in the trControl
matrix column.
}
