% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JBH_Misc.R
\name{graph_ma}
\alias{graph_ma}
\title{Plot correlation between simulated biomass and index using different moving
average parameters.

A ggplot is returned with the correlation for different moving average
parameters over range 0-1. Higher correlation is better. When the moving
average is zero, it will not change, whereas 1 indicates previous index
values are ignored. The moving average reduces the index noise. This
graph may not show a clear optimum, in which case a value around 0.5 is
recommended}
\usage{
graph_ma(jabba_fit)
}
\arguments{
\item{jabba_fit}{A list returned from package JABBA \code{fit_jabba} function}
}
\value{
A ggplot plotting moving average parameter against index/biomass
correlation
}
\description{
Plot correlation between simulated biomass and index using different moving
average parameters.

A ggplot is returned with the correlation for different moving average
parameters over range 0-1. Higher correlation is better. When the moving
average is zero, it will not change, whereas 1 indicates previous index
values are ignored. The moving average reduces the index noise. This
graph may not show a clear optimum, in which case a value around 0.5 is
recommended
}
