% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JBH_SimOutput.R
\name{graph_sim_BMSY_FMSY}
\alias{graph_sim_BMSY_FMSY}
\title{Plot the stock status and harvest rate for a single HCR test simulation.

A ggplot is returned with the biomass as a proportion of BMSY and H as a
proportion of HMSY over time.}
\usage{
graph_sim_BMSY_FMSY(HCR_sim, type = "both")
}
\arguments{
\item{HCR_sim}{Fishery HCR simulation result list from the MSE (see
\link{create_HCR_MSE}).}

\item{type}{Either "both", "bmsy" or "hmsy" to indicate which plots are
provided}
}
\value{
A probability density plot of biomass dynamics projection under a HCR
}
\description{
Plot the stock status and harvest rate for a single HCR test simulation.

A ggplot is returned with the biomass as a proportion of BMSY and H as a
proportion of HMSY over time.
}
