% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimOutput.R
\name{graph_sim_Btar_Ftar}
\alias{graph_sim_Btar_Ftar}
\title{Plot the stock status and harvest rate for a single HCR test simulation.}
\usage{
graph_sim_Btar_Ftar(HCR_sim, type = "both")
}
\arguments{
\item{HCR_sim}{Fishery HCR simulation result list from the MSE
(see \link{create_fishblicc_MSE}, \link{create_JABBA_MSE}, \link{create_ptStan_MSE}).}
}
\value{
A probability density plot of biomass dynamics projection under a HCR
}
\description{
A ggplot is returned with the biomass as a proportion of BMSY and H as a
proportion of HMSY over time.
}
