% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JBH_SimOutput.R
\name{graph_sim_catch}
\alias{graph_sim_catch}
\title{Plot the catches for a single HCR test simulation.}
\usage{
graph_sim_catch(HCR_sim)
}
\arguments{
\item{HCR_sim}{Fishery HCR simulation result list from the MSE
(see \link{create_HCR_MSE}).}
}
\value{
graph showing binned probability for observed and simulated catch
time series
}
\description{
Plot the catches for a single HCR test simulation.
}
