% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JABBA_Create.R
\name{JABBA_convert_jabba2list}
\alias{JABBA_convert_jabba2list}
\title{Does minimal checks on jabba fit and ensures it is in a list}
\usage{
JABBA_convert_jabba2list(jabba_fit)
}
\arguments{
\item{jabba_fit}{A list returned from package JABBA \code{fit_jabba} function}
}
\value{
jabba_fit in a list
}
\description{
Does minimal checks on jabba fit and ensures it is in a list
}
