% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JABBA_Create.R
\name{JABBA_MSY_refpt}
\alias{JABBA_MSY_refpt}
\title{Extracts MSY reference points from a JABBA model fit}
\usage{
JABBA_MSY_refpt(jabba_list, ref_year = NULL)
}
\arguments{
\item{jabba_list}{a single jabba fit or a list of jabba fits}

\item{ref_year}{Year to use for relative change in effort}
}
\value{
List of MSY reference points and a reference year for fishing
mortality for relative changes to that year
}
\description{
JABBA uses the harvest rate, H, which is catch / biomass and limited to the
range 0-1. For effort control, this is converted to a fishing mortality
equivalent (H=1-exp(-F)), which avoids problems with boundary conditions.
}
