% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jabba_seabob}
\alias{jabba_seabob}
\title{An example JABBA fit}
\format{
An object of class \code{list} of length 25.
}
\source{
CRFM 2025. Report of the CRFM Continental Shelf Fisheries
Working Group (CSWG): Intersessional Activities of April 2024 – March 2025 &
Consultation with Suriname Seabob Working Group and Seabob Fishery
Stakeholders of 04 April 2025. CRFM Fishery Report – 2025/1. 23p.
}
\usage{
jabba_seabob
}
\description{
The "data" list An list of fitted saved from a JABBA output from a Pella-Thomlinson model fitted to a small shrimp species (seabob) in Suriname.
}
\keyword{datasets}
