% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishblicc_Create.R
\name{meanLength_func}
\alias{meanLength_func}
\title{Creates a function to estimate the mean length from length frequency data
matrices}
\usage{
meanLength_func(blicc_ld, gears2use = "All")
}
\arguments{
\item{blicc_ld}{A list of settings used in fitted the length-based catch
curve fishblicc model}

\item{gears2use}{Either an integer vector of gears or "All"}
}
\value{
A function taking 1 parameter, a matrix list of length frequency data
}
\description{
The function is based on the fishblicc data list: blicc_ld. The resulting
function takes a list of matrices with the same dimensions, one for each
gear, with the columns aligned to the length bins and rows the number of MCMC
simulations.
}
