% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JBH_Create.R
\name{MSY_refpt}
\alias{MSY_refpt}
\title{Extracts MSY reference points from a JABBA model fit}
\usage{
MSY_refpt(jabba_fit, ref_year = NULL)
}
\arguments{
\item{jabba_fit}{A list returned from package JABBA \code{fit_jabba} function}

\item{ref_year}{Year to use for relative change in effort}
}
\value{
List of MSY reference points
}
\description{
JABBA uses the harvest rate, H, which is catch / biomass and limited to the
range 0-1. For effort control, this is converted to a fishing mortality
equivalent (H=1-exp(-F)), which avoids problems with boundary conditions.
}
