% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCRPerform.R
\name{past_performance}
\alias{past_performance}
\title{Calculate retrospective performance from fitted model}
\usage{
past_performance(HCR_MSE)
}
\arguments{
\item{HCR_MSE}{HCR function produced by \link{create_fishblicc_MSE},
\link{create_JABBA_MSE} or \link{create_ptStan_MSE}}
}
\value{
reference point list updated with past performance values
}
\description{
Calculate the performance measures for observed past performance for the
fitted model from the MSE function. See \link{HCR_performance} for interpretation
of the performance measures.
}
