% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JABBAstan_Create.R
\name{pt_create_linear_control}
\alias{pt_create_linear_control}
\title{Calculate a linear sliding control to be applied based on the index value}
\usage{
pt_create_linear_control(trIndex, trControl, change_limit = NA)
}
\arguments{
\item{trIndex}{A sorted matrix of index inflexion points from low to high.
Same size as \code{trControl}}

\item{trControl}{A sorted matrix of control inflection points defining linear
HCR.}

\item{change_limit}{An annual limit on the control change as a proportion}

\item{control_type}{Either "Effort" or "Catch" or "Both"}
}
\value{
A function to calculate the new control from the index (vector) and previous
year's controls (matrix)
}
\description{
The function creates a function to calculate a control based on the index,
previous control and HCR parameters held in the HCR list.
}
