% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JABBAstan_Create.R
\name{pt_create_mean_CPUE}
\alias{pt_create_mean_CPUE}
\title{Collapses multiple CPUE indices into single weighted geometric mean index}
\usage{
pt_create_mean_CPUE(stanfit)
}
\value{
A function taking a matrix with CPUE in columns and returning a
single vector index
}
\description{
A function is returned taking a matrix of CPUE with columns for each
CPUE index and convert to a single vector index. The indices are assumed to
be log-normally distributed, so the geometric mean is taken weighted by the
median JABBA observation error estimates (tau2).
}
