% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JABBAstan_Create.R
\name{pt_MSY_mean_refpt}
\alias{pt_MSY_mean_refpt}
\title{Extracts MSY reference points from a JABBAstan model fit}
\usage{
pt_MSY_mean_refpt(stan_fit)
}
\arguments{
\item{stan_fit}{A stanfit object}
}
\value{
List of MSY reference points for different variation
}
\description{
This only works on a single gear. To extend it, a function will need to be
provided to convert catch-effort to the HCR index.
}
