% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JBH_Create.R
\name{run_HCR_MSE}
\alias{run_HCR_MSE}
\title{Run an HCR MSE based on a table of parameters and HCR MSE function}
\usage{
run_HCR_MSE(HCR_df, HCR_MSE)
}
\arguments{
\item{HCR_df}{A data frame produced by eg \link{define_HCR_test_range} with
HCR parameters}

\item{HCR_MSE}{HCR function produced by \link{create_HCR_MSE}}
}
\value{
A data frame with the original parameters and additional columns
containing the performance indicators.
}
\description{
The function runs the HCR MSE based on the set of HCR parameters in the
dataframe and the provided MSE function. The performance indicators are
added as additional columns to the dataframe.
}
