% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JBH_Create.R
\name{run_HCR_MSE_list}
\alias{run_HCR_MSE_list}
\title{Run an HCR MSE based on a table of parameters and HCR MSE function over a
list of JABBA fits}
\usage{
run_HCR_MSE_list(
  jabba_list,
  HCR_df,
  control_type = "Effort",
  ProjLength = 50,
  nsim = 1000,
  ref_pt = standard_risk_ref_pt(),
  rho = 1
)
}
\arguments{
\item{jabba_list}{a single jabba fit or a list of jabba fits}

\item{HCR_df}{A data frame produced by eg \link{define_HCR_test_range} with
HCR parameters}

\item{control_type}{Type of control applied by HCR either "Effort" or "Catch"}

\item{ProjLength}{Projection length in years}

\item{nsim}{Number of simulations to run}

\item{ref_pt}{Specifies risk based reference points in a list. Otherwise
defaults apply.}

\item{rho}{Conversion factor from effort (f) to fishing mortality, so
H=(1-e(-rho*f)). Only used if control_type is "Effort". Control defaults to
fishing mortality (rho=1).}

\item{risk_rp}{Risk based reference levels defining targets relative to MSY
and maximum acceptable risks}
}
\value{
A data frame with the original parameters and additional columns
containing the performance indicators.
}
\description{
The function runs the HCR MSE based on the set of HCR parameters in the
dataframe and the provided MSE function on a list of JABBA fits and returns
the combined results in a dataframe. This is the same as \link{run_HCR_MSE}, but
instead of a single JABBA fit, the function works for multiple alternative
fits such as sensitivity analyses.
}
