% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunMSE.R
\name{run_HCR_MSE_para}
\alias{run_HCR_MSE_para}
\title{Parallel version of \link{run_HCR_MSE}}
\usage{
run_HCR_MSE_para(HCR_df, HCR_MSE)
}
\arguments{
\item{HCR_df}{A data frame produced by eg \link{define_HCR_test_range} with
HCR parameters}

\item{HCR_MSE}{HCR function produced by \link{create_fishblicc_MSE},
\link{create_JABBA_MSE} or \link{create_ptStan_MSE}}
}
\value{
A data frame with the original parameters and additional columns
containing the performance indicators.
}
\description{
The function runs the HCR MSE based on the set of HCR parameters in the
dataframe and the provided MSE function. The performance indicators are
added as additional columns to the dataframe. This should be faster than the
serial version when running several HCR. The function uses the package
\code{furrr}.
}
