% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishblicc_Create.R
\name{set_sim_recruit}
\alias{set_sim_recruit}
\title{Sets the recruitment deviates type, whether shared between simulations or
independent}
\usage{
set_sim_recruit(HCR_MSE, sim_recruit)
}
\arguments{
\item{HCR_MSE}{HCR_MSE function created by \link{create_fishblicc_MSE}}

\item{sim_recruit}{Whether simulated recruitment deviates are independent
for each simulation or shared across simulations}
}
\description{
Sets the recruitment deviates type, whether shared between simulations or
independent
}
