% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishblicc_Create.R
\name{SPR_func}
\alias{SPR_func}
\title{Creates a function to estimate the SPR from length frequency data matrices}
\usage{
SPR_func(blicc_ld, gears2use = "All")
}
\arguments{
\item{blicc_ld}{A list of settings used in fitted the length-based catch
curve fishblicc model}

\item{gears2use}{Either an integer vector of gears or "All"}
}
\description{
Creates a function to estimate the SPR from length frequency data matrices
}
