% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishblicc_Create.R
\name{SPR_onlyF_para}
\alias{SPR_onlyF_para}
\title{Creates a function to estimate the SPR from length frequency data matrices}
\usage{
SPR_onlyF_para(blicc_ld, gears2use = "All")
}
\arguments{
\item{blicc_ld}{A list of settings used in fitted the length-based catch
curve fishblicc model}

\item{gears2use}{Either an integer vector of gears or "All"}
}
\description{
Parallel version. Only estimates fishing mortality. Selectivity and other parameters are fixed to
the prior modes. Still too slow to use.
}
