% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JBH_Create.R
\name{combine_HCR_results}
\alias{combine_HCR_results}
\title{Combines multiple HCR simulation results into a single tibble}
\usage{
combine_HCR_results(HCR_res, HCR_df, HCR_MSE)
}
\arguments{
\item{HCR_res}{A list of results from \link{HCR_performance}}

\item{HCR_df}{A data frame produced by eg \link{define_HCR_test_range} with
HCR parameters}

\item{HCR_MSE}{HCR function produced by \link{create_HCR_MSE}}
}
\value{
A tibble containing combined results from the simulations
}
\description{
Combines multiple HCR simulation results into a single tibble
}
