% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JBH_Create.R
\name{create_HCR_MSE}
\alias{create_HCR_MSE}
\title{Create a HCR MSE function}
\usage{
create_HCR_MSE(
  jabba_fit,
  control_type = "Effort",
  ProjLength = 50,
  nsim = 1000,
  ref_pt = standard_risk_ref_pt(),
  rho = NULL
)
}
\arguments{
\item{jabba_fit}{A list returned from package JABBA \code{fit_jabba} function}

\item{control_type}{Type of control applied by HCR either "Effort" or "Catch"}

\item{ProjLength}{Projection length in years}

\item{nsim}{Number of simulations to run}

\item{ref_pt}{Specifies risk based reference points in a list. Otherwise
defaults apply.}

\item{rho}{Conversion factor from effort (f) to fishing mortality, so
H=(1-e(-rho*f)). Only used if control_type is "Effort"}
}
\value{
A function that will apply the HCR using the JABBA model fit and
specified HCR parameters: index-control inflection point vectors, the
control change limit and moving average parameter.
}
\description{
The HCR function returns simulation results from HCR parameters using a
jabba model fit results.
}
