% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JABBAstan_Create.R
\name{create_ptStan_MSE}
\alias{create_ptStan_MSE}
\title{The functions below implement a simplified harvest control rule projection
based on a fitted JABBA model. The method will work with some adaptation on
any application of a Pella-Thomlinson MCMC (or related) fitted model to generate projected stock
status over the projection period. It is configured for the Schaefer
model and the Pella-Thomlinson, so the m parameter is used.
There is no checking that parameters are correctly specified in this version
Workflow}
\usage{
create_ptStan_MSE(
  stan_fit,
  stan_dat,
  proj_length = 50,
  nsim = 1000,
  ref_pt = standard_risk_ref_pt()
)
}
\arguments{
\item{stan_fit}{A stanfit object}

\item{stan_dat}{Data list used to fit the Stan model}

\item{proj_length}{Projection length in years}

\item{nsim}{Number of simulations to run}

\item{ref_pt}{Specifies risk based reference points in a list. Otherwise
defaults apply.}

\item{control_type}{Type of control applied by HCR either "Effort" or "Catch"}
}
\value{
A function that will apply the HCR using the stan model fit and
specified HCR parameters: index-control inflection point vectors, the
control change limit and moving average parameter.
}
\description{
jabba_fits is a single fit or list of fits.
}
\details{
\enumerate{
\item Generate a jabba_fits
\item Define the linear HCR, ma, change_limit ranges to test and generate a tibble of these values
\item Run all HCR on JABBA fit info generating performance indicators
\item Generate plots of HCR performance with rejected HCR in red.
}

Create a HCR MSE function using a Stan fit of the JABBA

The HCR function returns simulation results from HCR parameters using a
jabba model fit results.
}
