% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCRPerform.R
\name{graph_HCR_catches}
\alias{graph_HCR_catches}
\title{Plot the average catch and mean catch range.

A ggplot is returned with the average catch and mean catch range on the
x- and y-axes respectively. Candidate HCR are plotted as black points and
rejected HCR as red points. A linear fit is shown going through the origin
indicating the exchange between increasing catch with increasing catch range.
HCR below this line indicate a better exchange rate between these two
measures.}
\usage{
graph_HCR_catches(HCR_df)
}
\arguments{
\item{HCR_df}{A data frame produced by eg \link{define_HCR_test_range} with
HCR parameters}
}
\value{
A ggplot object plotting the HCR performance catches
}
\description{
Plot the average catch and mean catch range.

A ggplot is returned with the average catch and mean catch range on the
x- and y-axes respectively. Candidate HCR are plotted as black points and
rejected HCR as red points. A linear fit is shown going through the origin
indicating the exchange between increasing catch with increasing catch range.
HCR below this line indicate a better exchange rate between these two
measures.
}
