% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JBH_HCRTest.R
\name{graph_HCR_decision}
\alias{graph_HCR_decision}
\title{Plot the HCR decision performance in relation to mistakes made by the
decision rule.

A ggplot is returned with the the proportion of Type 1 (false positive) and
Type 2 (false negative) errors. For Type 1, the decision rule results in
management intervention reducing harvest when the stock in reality is at or
above its target level. For Type 2 (false negative), the decision rule does
not harvest when the stock is in reality below the limit reference point,
which in this case is considered worse as it is less precautionary. Higher
error rates indicate poorer HCR performance and might be improved by
adjusting the index trigger points.

Candidate HCR are plotted as black points and rejected HCR as red points.
Risk based reference points are indicated by dotted vertical and horizontal
lines, which are set at default values if no alternative candidate
definition. is provided.}
\usage{
graph_HCR_decision(HCR_df, ref_pt = NULL)
}
\arguments{
\item{HCR_df}{A data frame produced by eg \link{define_HCR_test_range} with
HCR parameters}
}
\value{
A ggplot object plotting the HCR performance catches
}
\description{
Plot the HCR decision performance in relation to mistakes made by the
decision rule.

A ggplot is returned with the the proportion of Type 1 (false positive) and
Type 2 (false negative) errors. For Type 1, the decision rule results in
management intervention reducing harvest when the stock in reality is at or
above its target level. For Type 2 (false negative), the decision rule does
not harvest when the stock is in reality below the limit reference point,
which in this case is considered worse as it is less precautionary. Higher
error rates indicate poorer HCR performance and might be improved by
adjusting the index trigger points.

Candidate HCR are plotted as black points and rejected HCR as red points.
Risk based reference points are indicated by dotted vertical and horizontal
lines, which are set at default values if no alternative candidate
definition. is provided.
}
