% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JBH_HCRTest.R
\name{graph_HCR_status_catches}
\alias{graph_HCR_status_catches}
\title{Plot the HCR stock status relative to the BMSY (TRP) and Blim (LRP)

A ggplot is returned with the the proportion of the simulation time across
all runs for an HCR below the limit (x-axis) and around the target (y-axis).
Candidate HCR are plotted as black points and rejected HCR as red points.
Risk based reference points are indicated by dotted vertical and horizontal
lines, which are set at default values if no alternative candidate
definition. is provided.}
\usage{
graph_HCR_status_catches(HCR_df)
}
\arguments{
\item{HCR_df}{A data frame produced by eg \link{define_HCR_test_range} with
HCR parameters}
}
\value{
A ggplot object plotting the HCR performance catches
}
\description{
Plot the HCR stock status relative to the BMSY (TRP) and Blim (LRP)

A ggplot is returned with the the proportion of the simulation time across
all runs for an HCR below the limit (x-axis) and around the target (y-axis).
Candidate HCR are plotted as black points and rejected HCR as red points.
Risk based reference points are indicated by dotted vertical and horizontal
lines, which are set at default values if no alternative candidate
definition. is provided.
}
