% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JBH_HCRTest.R
\name{graph_linear_HCR}
\alias{graph_linear_HCR}
\title{Plot harvest control rules with the stock status index on the x-axis and
control on the y-axis.

A ggplot is returned with the HCR plotted as a line between
the index of stock status and control.}
\usage{
graph_linear_HCR(HCR_df, HCR_sim = NULL, HCR_ID = TRUE)
}
\arguments{
\item{HCR_df}{A data frame produced by eg \link{define_HCR_test_range} with
HCR parameters}

\item{HCR_sim}{Fishery HCR simulation result list from the MSE (see
\link{create_HCR_MSE}).}

\item{HCR_ID}{Logical - whether to include the HCR ID as a factor in the
plot. Will only work for 12 or fewer HCR.}
}
\value{
A ggplot object plotting the HCR's
}
\description{
Plot harvest control rules with the stock status index on the x-axis and
control on the y-axis.

A ggplot is returned with the HCR plotted as a line between
the index of stock status and control.
}
