% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JABBAstan_Create.R
\name{pt_create_HCR_index}
\alias{pt_create_HCR_index}
\title{Create the empirical abundance index update function.}
\usage{
pt_create_HCR_index(ma, func = NULL, se)
}
\arguments{
\item{ma}{Moving average parameter for index time series}

\item{func}{Function used to combine multiple CPUE series into a single index}

\item{se}{Observation standard error to be added in to the index}
}
\value{
A function taking a single CPUE list as a parameter that can be used
to calculate the HCR index
}
\description{
The index function created calculates a moving average from one or more
abundance indices, usually CPUE.
}
