% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCRPerform.R
\name{rank_HCR}
\alias{rank_HCR}
\title{Ranks HCR based on performance indicators and evaluates whether the HCR is a
candidate.

The resulting ranks and whether the HCR is a candidate or not is added as
columns to the data frame.}
\usage{
rank_HCR(HCR_df, HCR_MSE)
}
\arguments{
\item{HCR_df}{A data frame produced by eg \link{define_HCR_test_range} with
HCR parameters}

\item{HCR_MSE}{HCR function produced by \link{create_fishblicc_MSE},
\link{create_JABBA_MSE} or \link{create_ptStan_MSE}}
}
\value{
HCR data frame with added evaluation and ranks
}
\description{
Ranks are based on catch (10th percentile) and stock state (proportion of
the simulation spent around the target level - that spent below Blim).
}
