% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineHCR.R
\name{standard_risk_ref_pt}
\alias{standard_risk_ref_pt}
\title{Defines standard risk-based reference points}
\usage{
standard_risk_ref_pt(B_tar = 1)
}
\arguments{
\item{B_tar}{Target reference point relative to type}
}
\value{
List of reference points
}
\description{
This maintains the reference points definition in one place and can be
edited. The reference points are in a list and consist of:
\itemize{
\item type = Basis for biomass target. e.g. 'MSY' or 'SPR40'
\item B_tar = Biomass at the target as a proportion of the unexploited biomass
\item B_tar_range = accepted target range around the target biomass point
\item B_lim = limit reference point as a proportion of the unexploited biomass
\item mostly, likely, highly likely, high certainty = probabilistic definitions
of each word not currently used
\item max_risk = maximum acceptable risk for the candidate HCR selection.
Primarily used to exlcude HCR with this risk or greater of falling below the limit
reference point B_lim
}
}
