% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JBH_Create.R
\name{standard_risk_ref_pt}
\alias{standard_risk_ref_pt}
\title{Defines standard risk-based reference points.}
\usage{
standard_risk_ref_pt(m = 2)
}
\value{
List of reference points
}
\description{
This is purely to maintain reference points definition in one place.
Can be edited.
}
