% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JBH_HCRTest.R
\name{table_HCR_performance}
\alias{table_HCR_performance}
\title{Table the standard performance measures for a set of candidate HCRs}
\usage{
table_HCR_performance(HCR_df)
}
\arguments{
\item{HCR_df}{A data frame produced by eg \link{define_HCR_test_range} with
HCR parameters}
}
\value{
A \code{flextable} containing the HCR performance measures
}
\description{
Table the standard performance measures for a set of candidate HCRs
}
