% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JBH_SimOutput.R
\name{table_sim_decision}
\alias{table_sim_decision}
\title{Table the proportion of each state-response for HCR decisions}
\usage{
table_sim_decision(HCR_sim)
}
\arguments{
\item{HCR_sim}{Fishery HCR simulation result list from the MSE (see
\link{create_HCR_MSE}).}
}
\value{
flextable object containing the state-response proportions for the
HCR
}
\description{
The decisions compare stock status versus response, where an initial response
is indicated by the HCR index being below the highest trigger point in the
HCR or below the lowest trigger in response to being below Blim.
}
