% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimOutput.R
\name{table_sim_decision}
\alias{table_sim_decision}
\title{Table the proportion of each state-response for HCR decisions}
\usage{
table_sim_decision(HCR_sim)
}
\arguments{
\item{HCR_sim}{Fishery HCR simulation result list from the MSE
(see \link{create_fishblicc_MSE}, \link{create_JABBA_MSE}, \link{create_ptStan_MSE}).}
}
\value{
flextable object containing the state-response proportions for the
HCR
}
\description{
Table the proportion of each state-response for HCR decisions
}
