% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JBH_SimOutput.R
\name{table_sim_performance}
\alias{table_sim_performance}
\title{Table the standard performance measures for a HCR simulation}
\usage{
table_sim_performance(HCR_sim)
}
\arguments{
\item{HCR_sim}{Fishery HCR simulation result list from the MSE (see
\link{create_HCR_MSE}).}
}
\value{
\code{flextable} containing the HCR performance measures
}
\description{
Table the standard performance measures for a HCR simulation
}
