% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JBH_SimOutput.R
\name{table_sim_status}
\alias{table_sim_status}
\title{Table the simulation proportion of time that the stock spends in each status
interval.}
\usage{
table_sim_status(HCR_sim)
}
\arguments{
\item{HCR_sim}{Fishery HCR simulation result list from the MSE
(see \link{create_HCR_MSE}).}
}
\value{
\code{flextable} containing the simulations status proportions for the HCR
}
\description{
Table the simulation proportion of time that the stock spends in each status
interval.
}
